move include of multisolver_interface in a specific include directory
