+++ src/cpp/multisolver_interface/CMakeLists.txt
@@ -14,15 +14,11 @@
 # Conditionnal settings of solver source files
 # ---------------------------------------------------------------------------
 list(APPEND Solver_sources
-	${CMAKE_CURRENT_LIST_DIR}/Solver.h
-	${CMAKE_CURRENT_LIST_DIR}/SolverAbstract.h
-	${CMAKE_CURRENT_LIST_DIR}/SolverFactory.h
 	${CMAKE_CURRENT_LIST_DIR}/SolverFactory.cpp
 )
 # CPLEX
 IF( CPLEX )
 	list(APPEND Solver_sources
-		${CMAKE_CURRENT_LIST_DIR}/SolverCplex.h
 		${CMAKE_CURRENT_LIST_DIR}/SolverCplex.cpp
 	)
 ENDIF( CPLEX )
@@ -30,7 +26,6 @@ ENDIF( CPLEX )
 # XPRESS
 IF( XPRESS )
 	list(APPEND Solver_sources
-		${CMAKE_CURRENT_LIST_DIR}/SolverXpress.h
 		${CMAKE_CURRENT_LIST_DIR}/SolverXpress.cpp
 	)
 ENDIF( XPRESS )
@@ -38,9 +33,7 @@ ENDIF( XPRESS )
 #Clp - CBc
 IF(COIN_OR)
 	list(APPEND Solver_sources
-		${CMAKE_CURRENT_LIST_DIR}/SolverCbc.h
 		${CMAKE_CURRENT_LIST_DIR}/SolverCbc.cpp
-		${CMAKE_CURRENT_LIST_DIR}/SolverClp.h
 		${CMAKE_CURRENT_LIST_DIR}/SolverClp.cpp
 	)
 ENDIF(COIN_OR)
@@ -54,8 +47,10 @@ add_library (solvers STATIC
 )
 
 target_include_directories (solvers
-	PUBLIC
-		${CMAKE_CURRENT_SOURCE_DIR}
+		PUBLIC
+			${CMAKE_CURRENT_SOURCE_DIR}/include
+		PRIVATE
+			${CMAKE_CURRENT_SOURCE_DIR}
 )
 
 # CPLEX
