+++ src/cpp/multisolver_interface/SolverCbc.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "SolverAbstract.h"
+#include "multisolver_interface/SolverAbstract.h"
 #include "Cbc_C_Interface.h"
 #include "CbcModel.hpp"
 #include "OsiClpSolverInterface.hpp"
