+++ tests/cpp/solvers_interface/test_reading_problem.cpp
@@ -1,7 +1,7 @@
 #pragma once
 #include "catch2.hpp"
 #include <iostream>
-#include "Solver.h"
+#include "multisolver_interface/Solver.h"
 #include "define_datas.hpp"
 
 TEST_CASE("Un objet solveur peut etre cree et detruit", "[read][init]") {
